extract schedulewriter and formatter to generator.py
